rebase on master, fix tests, cleanup
